+++ src/Geometry.jl
@@ -197,13 +197,23 @@ struct DistributedCartesianDescriptor{A,B,C}
   descriptor::C
   function DistributedCartesianDescriptor(
     ranks::AbstractArray{<:Integer},
-    mesh_partition :: NTuple{Dc,<:Integer},
-    descriptor :: CartesianDescriptor{Dc}
+    mesh_partition::NTuple{Dc,<:Integer},
+    descriptor::CartesianDescriptor{Dc}
   ) where Dc
     A = typeof(ranks)
     B = typeof(mesh_partition)
     C = typeof(descriptor)
     new{A,B,C}(ranks,mesh_partition,descriptor)
+  end
+end
+
+function Base.show(io::IO,k::MIME"text/plain",desc::DistributedCartesianDescriptor)
+  ranks = desc.ranks
+  map_main(ranks) do r
+    nranks = desc.mesh_partition
+    ncells = desc.descriptor.partition
+    f(x) = join(x,"x")
+    print(io,"$(f(ncells)) CartesianDescriptor distributed in $(f(nranks)) ranks")
   end
 end
 
@@ -413,6 +423,26 @@ function Geometry.DiscreteModel(
   end
 
   GenericDistributedDiscreteModel(models,gids)
+end
+
+# UnstructuredDiscreteModel
+
+const DistributedUnstructuredDiscreteModel{Dc,Dp,A,B,C} = 
+  GenericDistributedDiscreteModel{Dc,Dp,<:AbstractArray{<:UnstructuredDiscreteModel},B,C}
+
+function Geometry.UnstructuredDiscreteModel(model::GenericDistributedDiscreteModel)
+  return GenericDistributedDiscreteModel(
+    map(UnstructuredDiscreteModel,local_views(model)),
+    get_cell_gids(model),
+  )
+end
+
+# Simplexify
+
+function Geometry.simplexify(model::DistributedDiscreteModel)
+  _model = UnstructuredDiscreteModel(model)
+  ref_model = refine(_model, refinement_method = "simplexify")
+  return UnstructuredDiscreteModel(Adaptivity.get_model(ref_model))
 end
 
 # Triangulation