+++ test/TestApp/src/TestApp.jl
@@ -12,7 +12,8 @@ module TestApp
   include("../../HeatEquationTests.jl")
   include("../../StokesOpenBoundaryTests.jl")
   include("../../AdaptivityTests.jl")
-  include("../../CartesianAdaptivityTests.jl")
+  include("../../AdaptivityCartesianTests.jl")
+  include("../../AdaptivityUnstructuredTests.jl")
   include("../../AdaptivityMultiFieldTests.jl")
   include("../../BlockSparseMatrixAssemblersTests.jl")
 end