+++ test/sequential/AdaptivityTests.jl
@@ -2,9 +2,11 @@ module AdaptivityTestsSeq
 
 using PartitionedArrays
 
-include("../CartesianAdaptivityTests.jl")
+include("../AdaptivityCartesianTests.jl")
+include("../AdaptivityUnstructuredTests.jl")
 with_debug() do distribute
-  CartesianAdaptivityTests.main(distribute)
-end 
+  AdaptivityCartesianTests.main(distribute)
+  AdaptivityUnstructuredTests.main(distribute)
+end
 
 end # module